Enable custom configuration for iOS maps
This allows setting up the library to work with Google Maps on iOS via
`react-native link` instead of only via CocoaPods.

Because Google-Maps-iOS-Utils is not distributed as a binary (at this
time), it disables functionality that depends on this library and raises
an exception that points to the reason why there's a problem.

Additionally:

- A missing file has been added back into the AirMaps.xcodeproj project.